3-connect-to-snowflake/12-metadata-viewer/requirements.txt
snowflake-connector-python
snowflake-snowpark-python==1.10.0
streamlit